Speed up GraphMol/Chirality.cpp/iterateCIPRanks
The changes in this patch increases throughput of iterateCIPRanks by
about 61.8%.  This patch consists of many changes that all combine to
contribute to the performance gains, and this is a list of the major
changes:

* Introduce Bond.getTwiceBondType to avoid floating point operations in
`floor(2*bond->getBondTypeAsDouble() + .1)`
* In RDGeneral/Ranking.h/rankVect, just store the last index (instead of
the last vector) and avoid creating a copy at the current index.
* In GraphMol/Chirality.cpp/iterateCIPRanks, reserve more capacity ahead
of time.
* In GraphMol/Chirality.cpp/iterateCIPRanks, avoid materializing
localEntry and then copying; instead, track the data to be added in a
`counts` vector and a `updated_nbr_idxs` vector.

In GraphMol/Chirality.cpp/iterateCIPRanks, we also make several misc
micro-optimizations that were individually measured to give a small
performance improvement.